requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
dill==0.3.8
faiss-cpu==1.8.0.post1
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.28
langsmith==0.1.98
MarkupSafe==2.1.5
mistralai==1.0.1
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.0
pydantic==2.8.2
pydantic_core==2.20.1
pyTelegramBotAPI==4.22.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
telebot==0.0.5
tenacity==8.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
python-dotenv==0.19.1